+++ R/convert.R
@@ -15,12 +15,10 @@
 #' These functions are intended for complex S3 objects from other packages.
 #' See [cal()] for a more generic constructor, e.g. using a data frame.
 #'
-#' @param x  object to be converted to a `cal` object.
+#' @param x  Object from another package to be converted to a `cal` object.
 #'
 #' @returns
-#' `cal` object: a data frame with two columns, `year` and `p`, representing
-#' the calibrated probability distribution. All other values are stored as
-#' attributes and can be accessed with [cal_metadata()].
+#' Vector of class `c14_cal` ([cal]).
 #'
 #' @family cal class methods
 #' @family c14 conversion functions
@@ -35,24 +33,12 @@ as_cal.CalDates <- function(x) {
     warning("calMatrix object in CalDates ignored.")
   }
 
-  caldates <- x$metadata
-  caldates$calGrid <- x$grids
+  caldates <- cal_recode_metadata(x$metadata, "CalDates", "cal")
+  caldates$pd <- purrr::map(x$grids, as_calp)
+  caldates$engine <- "rcarbon"
+  caldates$engine_version <- packageVersion("rcarbon")
 
-  # TODO: automate attribute recoding via cal_recode_metadata()
-  purrr::pmap(caldates, ~with(list(...),
-                              new_cal(calGrid,
-                                      era = "cal BP",
-                                      lab_id = DateID,
-                                      cra = CRA,
-                                      error = Error,
-                                      curve = CalCurve,
-                                      reservoir_offset = ResOffsets,
-                                      reservoir_offset_error = ResErrors,
-                                      calibration_range = c(StartBP, EndBP),
-                                      normalised = Normalised,
-                                      F14C = F14C,
-                                      p_cutoff = CalEPS
-                              )))
+  do.call(new_cal, caldates)
 }
 
 #' Convert cal objects to an rcarbon CalDates object