+++ NEWS.md
@@ -1,7 +1,10 @@
 # c14 development version
 
-* Revised `cal` class, representing calibrated radiocarbon dates:
-  * Probability distributions are now represented by `calp` class.
+* Revised the `cal` class, representing calibrated radiocarbon dates:
+  * `cal` is now a `vctrs::rcrd` subclass with fields for metadata and the probability distribution (see below)
+  * Metadata specification has been changed.
+* Probability distributions are now represented by `calp`, a data frame subclass.
+  * The `age` column of a `calp` object is an `era::yr` vector, removing the need to specify the epoch used as part of the `cal` metadata.
   * `as_calp()` converts objects from other packages (`rcarbon::calGrid`) to `calp`.
 * Added summary functions for calibrated dates (`cal` class):
   * `c14_point()` for point estimates